Moved Navigator to navigation package
