+++ src/main/g8/test/navigation/FakeNavigator.scala
@@ -1,8 +1,9 @@
-package utils
+package navigation
 
 import play.api.mvc.Call
 import pages._
 import models.{Mode, NormalMode}
+import utils.UserAnswers
 
 class FakeNavigator(desiredRoute: Call, mode: Mode = NormalMode) extends Navigator {
   override def nextPage(page: Page, mode: Mode): UserAnswers => Call = _ => desiredRoute
