+++ src/main/g8/test/navigation/NavigatorSpec.scala
@@ -1,4 +1,4 @@
-package utils
+package navigation
 
 import base.SpecBase
 import org.mockito.Mockito._
@@ -6,6 +6,7 @@ import org.scalatest.mockito.MockitoSugar
 import controllers.routes
 import pages._
 import models._
+import utils.UserAnswers
 
 class NavigatorSpec extends SpecBase with MockitoSugar {
