+++ src/main/scaffolds/stringPage/generated-test/controllers/$className$ControllerSpec.scala
@@ -3,7 +3,7 @@ package controllers
 import play.api.data.Form
 import play.api.libs.json.JsString
 import uk.gov.hmrc.http.cache.client.CacheMap
-import utils.FakeNavigator
+import navigation.FakeNavigator
 import connectors.FakeDataCacheConnector
 import controllers.actions._
 import play.api.test.Helpers._
